reference/e2e/pyproject.toml
[project]
name = "ttt-e2e"
version = "0.1.0"
description = "End-to-End Test-Time Training for Long Context"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "chex>=0.1.87",
  "einops>=0.8.0",
  "einx>=0.3.0",
  "equinox>=0.11.12",
  "grain>=0.2.7",
  "h5py>=3.12.1",
  "humanfriendly>=10.0",
  "hydra-core>=1.3.2",
  "hydra-submitit-launcher>=1.2.0",
  "jax[cuda12]<0.6",
  "jaxtyping>=0.2.36",
  "matplotlib>=3.9",
  "numcodecs>=0.15.1",
  "numpy>=2.0",
  "optax>=0.2.4",
  "orbax-checkpoint>=0.11.13",
  "pytest>=8.3.5",
  "tensorstore>=0.1.71",
  "tqdm>=4.67.1",
  "wandb>=0.19",
  "zarr>=3.0.4",
]

[project.scripts]
train = "ttt.train:main"

[tool.setuptools.packages.find]
include = ["ttt", "hydra_plugins"]

[tool.uv]
package = true
default-groups = ["dev"]

[dependency-groups]
dev = ["ruff>=0.8"]

[tool.ruff]
line-length = 160

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "UP", "I", "ARG"]
ignore = [
  "E731",  # Do not assign a lambda expression, use a def
  "F722",  # Errors for jaxtyping syntax
  "E722",  # Do not use bare 'except'
  "E741",  # Ambiguous variable name
]
